Remove "Skipping the check of repos" from report (#986)
The "Skipping the check of repositories due to the use of RHSM for the
conversion" informational message is confusing in the pre-conversion report
as it does not have the context it has in the terminal where it's pre-empted
by "TASK - [Prepare: Check if --enablerepo repositories are accessible]".

This message woulld add no value for the user even if the context is
added. Hence the removal.