Make report msgs about removed pkgs less confusing
Currently the pre-conversion analysis report states that certain
packages were removed. But by the end of the conversion a rollback
happens and the packages are installed back. So at the time of reading
the report the packages are no longer removed.

This change rewords the messages to state that the packages are to be
removed during the conversion.