[RHELC-1252] Fix sub-man-rhsm-certs installation on 8.5/8.8 (#998)
The https://github.com/oamg/convert2rhel/pull/991 was dealing with a
convert2rhel failure stemming from two different architectures of the
subscription-manager-rhsm-certificates package being available in UBI
repos. That fix however caused convert2rhel failures on older system minor
versions where only x86_64 arch was available.

In the meantime Pino Toscano asked to have the UBI repos fixed in a way
that there is now just one architecture of the package. With that we
don't need to hardcode a specific arch in convert2rhel.