Fix yumdownloader parsing with yum-plugin-local
When the yum-plugin-local is installed, the yumdownloader output prints
information about building a local copy of a repository. It prints it
after the downloaded rpm filename and until now we expected the rpm
filename to be on the last line of the yumdownloader output.

This change looks for the rpm name on any line of the yumdownloader
output, not just the last line. Because we always download a single
package, there should be just one rpm filename present in the output.

This commit also makes sure that whenever we can't find the downloaded
rpm filename in the yumdownloader call output, we stop the conversion
because continuing would lead to an incomplete rollback.