[RHELC-1272] Fix a rollback failure with yum-plugin-local installed (#1001)
* Fix yumdownloader parsing with yum-plugin-local

When the yum-plugin-local is installed, the yumdownloader output prints
information about building a local copy of a repository. It prints it
after the downloaded rpm filename and until now we expected the rpm
filename to be on the last line of the yumdownloader output.

This change looks for the rpm name on any line of the yumdownloader
output, not just the last line. Because we always download a single
package, there should be just one rpm filename present in the output.

This commit also makes sure that whenever we can't find the downloaded
rpm filename in the yumdownloader call output, we stop the conversion
because continuing would lead to an incomplete rollback.

* yum-plugin-local: Add an integration test

Reproducer of the issue:
1. Install yum-plugin-local package and centos-logos on a CentOS-7.9
system
2. Run `convert2rhel analyze` and wait for it perform a rollback
3. The system is missing centos-logos

The new test installs a *-logos package, runs convert2rhel analyze, and
makes sure the *-logos package is installed after the rollback.

* yum-plugin-local: apply test review suggestions

* Fix parsing "using local copy" in pkg download

* Modify the test

* run the convert2rhel two times to verify the scenario works for both -
* packages got downloaded and packages are already in the backup dir
* add fixture to clear the downloaded files

Signed-off-by: Daniel Diblik <ddiblik@redhat.com>

---------

Signed-off-by: Daniel Diblik <ddiblik@redhat.com>
Co-authored-by: Daniel Diblik <ddiblik@redhat.com>