Refactor files backup
Refactoring the way we organize the files backup related code.
Currently, we just throw everything inside the backup.py module and call
the classes/functions from there.

With this commit, we refactor have it's own module inside the backup
folder. The idea is to bring all of the others backup codes to it's own
module as well, and leave the backup/__init__.py to be a base module
that will have only base code for the other modules.

Signed-off-by: Rodolfo Olivieri <rolivier@redhat.com>