Fix file clash in NewRestorableFile
There was a clash happening in the way we backup files, as if there is a
file to backup that has the same name as one directory or another file
already created, an error will be thrown to the user.

This commit fixes that clash by generating an hash directory based on
the dirname of the absolute path.

Signed-off-by: Rodolfo Olivieri <rolivier@redhat.com>