Apply suggestions from review
This commit prevents the RestorableFile being enabled if the target
filepath is not a file.

It also removes duplicated tests cases where the test was the same, but
using the backup controller instead. It is not required to use backup
controller instance for those type of tests, since we want to test the
RestorableFile class itself, not it's integration with the controller.

Signed-off-by: Rodolfo Olivieri <rolivier@redhat.com>