[RHELC-1375] Port RestorableFile to BackupController (#1045)
* Port RestorableFile to BackupController

First part of the port for the remaining items for the BackupController.

Changes included in this work are:

1. Remove the old RestorableFile class in order to use the
   NewRestorableFile
2. Rename the NewRestorableFile back to RestorableFile
3. Fix a bug where test pollution was happening in the BackupController
   global object

Signed-off-by: Rodolfo Olivieri <rolivier@redhat.com>

* Rename test reference

Signed-off-by: Rodolfo Olivieri <rolivier@redhat.com>

* Fix CodeQL warnings

Signed-off-by: Rodolfo Olivieri <rolivier@redhat.com>

* Refactor files backup

Refactoring the way we organize the files backup related code.
Currently, we just throw everything inside the backup.py module and call
the classes/functions from there.

With this commit, we refactor have it's own module inside the backup
folder. The idea is to bring all of the others backup codes to it's own
module as well, and leave the backup/__init__.py to be a base module
that will have only base code for the other modules.

Signed-off-by: Rodolfo Olivieri <rolivier@redhat.com>

* Apply review suggestions

Signed-off-by: Rodolfo Olivieri <rolivier@redhat.com>

* Apply suggestions from review

This commit prevents the RestorableFile being enabled if the target
filepath is not a file.

It also removes duplicated tests cases where the test was the same, but
using the backup controller instead. It is not required to use backup
controller instance for those type of tests, since we want to test the
RestorableFile class itself, not it's integration with the controller.

Signed-off-by: Rodolfo Olivieri <rolivier@redhat.com>

---------

Signed-off-by: Rodolfo Olivieri <rolivier@redhat.com>