Refactor Removed Packages classes in handle_packages
Prior to this refactor, the restore was acting as a single rpm
transaction to restore the removed packages, the problem is that the new
RestorablePackage now deals with a instantiation of those removals and
was causing problems during the rollback to treat each package as a
separate rollback.

This commit is introducing a refactor to allow both of them to happen at
the same time, avoiding conflicts.

Signed-off-by: Rodolfo Olivieri <rolivier@redhat.com>