Update yum transaction logic
Split the run_transaction method to have two variants, one that will be
executed without being in a subprocess, and the other one that will
execute the yum operations under a subprocess. This is necessary so we
can keep the backup/restore using the RestorablePackage class consistent
and working without modifying other frameworks.

Signed-off-by: Rodolfo Olivieri <rolivier@redhat.com>