Add API documentation.
This patch adds the initial setup if the Python Sphinx Autodocs.

We use autosummary extension to generate recursively the tree of all
modules, classes and functions. Then we additionally add two pages
(actions.rst and pkgmanager.rst) to show the possibility to add custom
content to the autogenerated documentation.

To build the docs you need

python3-sphinx
python3-sphinx-autodoc-typehints

installed on the environment (via dnf or pip)