Add auto retry for coverage analysis
The "Upload coverage to Codecov" step of the coverage analysis often
fails due to network issues with an error like:
```
Commit creating failed: {"error": "Server Error (500)"}
```

GitHub coverage action doesn't have a retry parameter, thus we use a
trick:

we add one more upload step exactly as the previous one, but we would
trigger it only in case the previous step failed.

Additionally, as retrying one upload too close to the other still gets a
failure we add a sleep step to wait for a minute before retrying.