Modify pre-commit to use latest pylint version
Update pre-commit config to use the latest version of pylint.
Previously, we pinned the version of pylint as we still had to support
older python versions because of RHEL 6.

Since we don't have this requirement anymore, we can update pylint and
only ignore the flags that we know that won't be available for Python
2.7.