Update call_yum_cmd to use reposdir and other opts
The call_yum_cmd function had to be updated to use reposdir, varsdir and
custom releasever because now the removal of the excluded/repofile
packages are being removing in a single action. This removal had to be
done in order to maintain an easier state for rollbacking the packages
removed by those two groups.