refactor: move `utils.py` to a module
As we kept using `utils.py` file and adding more and more things it
became quite bloated. To solve this the best thing to do, while also
making it easy to test and reduce cyclic imports is to split it into
smaller parts.

This takes care of the initial move by putting everything on the module
level while still having backwards compatibility for the time being.

To make this easier for ourselves we should migrate the file into
more modules over time