[RHELC-1382] Port RestorablePackage and ChangedRPMPackagesController (#1091)
* Port RestorablePackage and ChangedRPMPackagesController

Merging the two classes together and working only with
RestorablePackage.

In this commit, we are introducing the refactor, as well as some other
minor refactors that were required to avoid circular dependency imports.

Signed-off-by: Rodolfo Olivieri <rolivier@redhat.com>

* Refactor Removed Packages classes in handle_packages

Prior to this refactor, the restore was acting as a single rpm
transaction to restore the removed packages, the problem is that the new
RestorablePackage now deals with a instantiation of those removals and
was causing problems during the rollback to treat each package as a
separate rollback.

This commit is introducing a refactor to allow both of them to happen at
the same time, avoiding conflicts.

Signed-off-by: Rodolfo Olivieri <rolivier@redhat.com>

* Apply suggestions from code review

Co-authored-by: Adam Hosek <hosek.adam@outlook.com>

* Update yum transaction logic

Split the run_transaction method to have two variants, one that will be
executed without being in a subprocess, and the other one that will
execute the yum operations under a subprocess. This is necessary so we
can keep the backup/restore using the RestorablePackage class consistent
and working without modifying other frameworks.

Signed-off-by: Rodolfo Olivieri <rolivier@redhat.com>

* Update convert2rhel/pkgmanager/handlers/yum/__init__.py

Co-authored-by: Adam Hosek <hosek.adam@outlook.com>

* Remove reposdir if not internet access is available

The original behavior does not contain the reposdir variable set, only
varsdir.

Signed-off-by: Rodolfo Olivieri <rolivier@redhat.com>

* Apply suggestions from code review

Signed-off-by: Rodolfo Olivieri <rolivier@redhat.com>

* Update call_yum_cmd to use reposdir and other opts

The call_yum_cmd function had to be updated to use reposdir, varsdir and
custom releasever because now the removal of the excluded/repofile
packages are being removing in a single action. This removal had to be
done in order to maintain an easier state for rollbacking the packages
removed by those two groups.

---------

Signed-off-by: Rodolfo Olivieri <rolivier@redhat.com>
Co-authored-by: Adam Hosek <hosek.adam@outlook.com>