fix: errors in utils.py after full test suite
Since of merging #1029 we get an issue with unit tests that causes a lot
of strange errors to occur. Likely due to a stream not being closed
correctly.

This aims to either fix or mitigate this from occuring during test runs
as it should not appear in convert2rhel by itself.