Add post conversion stages for Action Framework
In this commit, we are introducing two new stages for enabling the post
conversion in the action framework.

We are also introducing an environment variable to not break the current
workflow and let the conversion function as-is for now. Anyone that is
willing to test and run the post conversion actions will need to use
this environment var.