Breaking change: Rename incomplete rollback envvar
We are moving away from using the word UNSUPPORTED in environment
variables as it is confusing and not very descriptive of what is
unsupported. Since this relates to the tool itself not supporting it in
a predictable way, we are renaming mentions of the word unsupported.

This change has to do with renaming
CONVERT2RHEL_UNSUPPORTED INCOMPLETE_ROLLBACK
to
CONVERT2RHEL_INCOMPLETE_ROLLBACK

CONVERT2RHEL_INCOMPLETE_ROLLBACK means that we can garantee a rollback
will be successful on a potential failure and user would have to specify
the environment variable themselves to understand this to not blame the
tool.

This change does not do anything else and no backwards compatibility is
considered as it is intended to be changed in a major version change.

Related
RHELC-1430