Install subscription-manager using package manager
In this commit, we are introducing the installation of
subscription-manager directly through a package manager, instead of
downloading the package first and then installing it with system package
manager.

This is necessary as we want to do GPG verification when installing
those set of packages, in which case, it was not available when we used
to download the packages and then install it via local install.

Notable changes in this commit are:
* Replacing installing subscription-manager to install it directly from
  the client-tools repository with gpg check enabled for all supported
  versions
* Minor refactor to RestorablePackageSet class to allow it to be generic
  and use custom repository paths and content to install packages
* Small tweak to call_yum_cmd to use the reposdir option as a list
  instead of a string
* Move subscription-manager constants to subscription.py module instead
  of the packages.py backup module