[RHELC-1465] Set consistent task prefix for pre-conversion (#1154)
Normalize task header for pre conversion

Some task headers for pre-conversion were using "Convert" instead of
"Prepare". Since we are dealing with actions that are present before we
start the conversion, all of those log tasks were changed to contain the
word "Prepare" instead to not sound confusing to the user as what is
happening.