chore: build RPMs on tags
This adds another trigger for the rpm builds so that we will be able to
attach rpms to specific tags. This primarily helps out with the weekly
tags we do on Fridays where we can easily link to it for the latest
RPM that should be decently stable

Note that these RPMs are still not extensively verified like we would
for an actual release. Instead they are weekly snapshots of the current
state of the repository.