Disable RHEL repos during checks
Related issue: [RHELC-884]

We need to disable RHEL repos when performing checks. If user registers the system to
satellite which provides Centos and RHEL repos and the RHEL repos would be enabled, it
would cause a fail. The recommended way is to disable the rhel repos prior the conversion,
but is expected it might be forgotten. This helps avoiding the problems caused by
wrong setup of the repos.

RHEL repos are disabled when:
* trying to download backup of package present on original system
* when checking if kernel and packages are up-to-date
* when retrieving repository info from packages (RemoveSpecialPkgs and
  ListThirdParty packages)