Drop internet connection check and its dependencies
In this commit, we are dropping the internet connection check and its
dependencies as well.

Some of them are not necessary, as we really want to use the system
repos backed up during the conversion, and not some hardcoded
repositories that we provide.

Historically, we used that for eus conversions, but right now it is not
necessary anymore.

The internet connection check brought a lot of difficulties in the past
and several bug reports. With this in mind, we decided to drop the
check, and assume the system will be configured in the correct way.