[RHELC-676] Unify parsing of the system release string (#897)
* Generic parser for system release content

There are several places where the system release string is parsed via regular expression. These expressions are different and adjusting them requires refactoring in multiple places.

This change introduces one parsing function which does a full match of the system-release string and reads all possible data points from it, and provides the result as a dictionary for later use by other tools.

The function replaces three other helper methods: _get_system_name, _get_system_version and _get_system_distribution_id, which are removed.

NOTE1: The parsing function could be a static method or even implemented as a utility function outside the class. The only reason why it is in the class is the logger object.

NOTE2: the function overlaps with the get_system_release_info function. We should check if this needs to be further unified.

* Rename match to matched to avoid name collision

* Ignote type hint on re.match

Co-authored-by: Freya Gustavsson <freya@spytec.se>

---------

Co-authored-by: Freya Gustavsson <freya@spytec.se>