Refactor call_yum_cmd to accept setopts only
Dropped the other two setopts we had defined in the past, namely,
varsdir and reposdir, in order to accept only the params through
setopts.

Since we are introducing setopts in the previous commit, it makes sense
to drop the varsdir and reposdir in favour of using only setopts, which
maps to the same interface for yum/dnf.