[RHELC-1497, RHELC-1499] Drop internet connection check and its dependencies (#1181)
* Drop internet connection check and its dependencies

In this commit, we are dropping the internet connection check and its
dependencies as well.

Some of them are not necessary, as we really want to use the system
repos backed up during the conversion, and not some hardcoded
repositories that we provide.

Historically, we used that for eus conversions, but right now it is not
necessary anymore.

The internet connection check brought a lot of difficulties in the past
and several bug reports. With this in mind, we decided to drop the
check, and assume the system will be configured in the correct way.

* Remove unused internet connection check tests

* Drop fixtures from pytest.ini

* Remove reposdir from is_loaded_kernel_latest action

At the point of execution, we don't have any repositories backed up, so
we don't need to try to use the reposdir.

* Check if reposdir is empty or does not exist

In case it is empty or does not exist, we should just not pass it to the
download_pkg function. This will prevent the download to fail because we
are using something that does not exist.

* Update convert2rhel/actions/system_checks/is_loaded_kernel_latest.py

* Remove hardcoded centos8 repos from "remove_repositories" fixture

The hardcoded repos are being removed in this PR

---------

Co-authored-by: Martin 'kokesak' Litwora <mlitwora@redhat.com>