fix: cornercase with userpass and keyorg
If a user provided username and password CLI arguments and had set
activation key and organization in the config file, it would fail to
execute convert2rhel. Same also happens if organization is set and
activation key isn't.

This is addressed by also unsetting organization if we detect username
and password.

Fixes RHELC-1525