Restore disabled repos during conversion in rollback
In a satellite analysis/conversion, we are disabling all repositories to
not interfer with the rest of the execution.

With this commit, we are introducing a way of checking what repositories
are enabled prior to disable them, and back the repositories names up.
In the rollback phase, we enable the repositories back on the system to
leave everything the way it was.