fix: require username and password to be specified (#1209)
We didn't require username and password to be specified when one was
provided but not the other. Such as username being provided but without
a password, or vice versa. Due to that convert2rhel fails much later
in the process than other warnings.

This addresses that with warning messages and aligns with other cases
like activation_key and org