Fix regression in installing removed packages during rollback
We were passing a set object to iterate over to compose a message of
packages that got removed during yum transaction. This problem was not
happening before as we were dealing with a success case of restoring
packages. In this regression, we found out that a set object does not
have indexing, therefore, thus bug was hidden from most of our tests and
pipeline.