Add exclude setopt to remaining repoquery calls
We need to exclude any packages set in the /etc/yum.conf set by the user
so we can make sure that when we run repoquery, we are really gathering
all the data we want from the command. This patch introduces the
overridable --setopt=exclude= in the remaining repoquery calls.