Remove RHSM repositories enablement fallback
In case of repositories not being able to be enabled, we won't fallback
to the normal RHSM repoids anymore. Instead, we want to throw an error
and tell the user what was the reason on why we couldn't enable the
repository.

This simplify the function that handles the enablement, making it much
more simpler now.