fix: prevent duplicate backups
It was possible for multiple packages to depend on the same file. As we
fetch the files for each package that needs to be backed up this caused
issues when rolling back a system as we try to restore it twice.

This fixes that issue by making sure we don't backup to the same path
multiple times

Signed-off-by: Freya Gustavsson <freya@venefilyn.se>