Require yum on RHEL 8+ based systems
The package manager on RHEL 8+ is dnf but for backwards compatibility
reason a yum rpm is available to create symlinks pointing to dnf.

If the yum package is missing, convert2rhel fails because it calls yum.

Having yum as a dependency of convert2rhel on RHEL 8+ based systems will
ensure that yum will be available (unless someone intentionally removes
yum after the convert2rhel installation but then ¯\_(ツ)_/¯).