Use no_rhsm to detect if enablerepos should be disabled
We have a special function that returns a list of repositories that
needs to be disabled during various command calls, and that list is
returning always the user enablerepos command line switch in it. This
patch only add the enablerepos option to the list when the --no-rhsm
switch is also applied.