Pick correct report results after inhibitors
There are two places where we raise the _InhibitorsFound exception, and
while this works correctly, we actually are reporting the wrong type of
data in one of them. We always assume that the type of data that will be
shown to the user is for pre-conversion, but in a further case down the
road of conversion, we could also show post conversion results.

This commit patches this bug and tries to correctly show the appropriate
result depending on the conversion state.