Add dependency for non redhat pkgs listing
Without this dependency, the action can run before the package
conversion in the system. Ideally, we want this to run after the
`KERNEL_PACKAGES_INSTALLATION`, as it is no use to have this run before the
package replacement.

Pointed out by @hosekadam in https://github.com/oamg/convert2rhel/pull/1256#discussion_r1718100982