Update code reference about toolopts and cli on codebase
This patch introduces the updated references to both toolopts and cli
throughout the codebase as some of the things we had before could be
changed to the new format, especially because no code get executed
before the CLI class parse all the options necessary.

We are not safe to call tool_opts even if nothing is populated inside
the class. Beware that it will still fail if we try to access any
attributes before they get created, but that should be a implementation
problem, not really a workflow one, since we expect that the CLI get
executed very early in the process.