chore: Simplify tests and rpmbuild generation
rpmbuilds are now handled in the same container as we run tests through
which helps us with maintaining less files. On top of that we can build
rpms through the published images we have on ghcr

On top of this we can now run tests in parallel through `make -j3 tests`
that will run all containers at once. This means we get results faster
at the expense of the output looking worse, but that is something we can
handle later

Future refactor of this could be to use compose files or kubes instead.
But as of right now this is plenty good for just a quick sanity check.