Add deprecation notice about environment variables
With the most recent work to migrate all environment variables to the
configuration file, we need now to warn the user that the variables will
be deprecated in order to use configuration file as our main place to
track configuarations.

This patch is introducing a very lightweight change to make sure that we
will be able modify the code in the future to remove all environment
variables.