refactor: Improved logger usage (#1373)
* refactor: move codebase to new logging syntax

Signed-off-by: Freya Gustavsson <freya@venefilyn.se>

* refactor: logging stages

This refactors the logging stages that we output when we log in specific
files depending on what we are doing. Meaning that any changes that
has to do with preparing or rollback will now output the stage within
the logger formatter rather than leaving it up to each class

E.g. any logs with "Prepare: text" or "Rollback: text" will now be
handled at the log-level

Signed-off-by: Freya Gustavsson <freya@venefilyn.se>

* Change format to use new stages style

Signed-off-by: Freya Gustavsson <freya@venefilyn.se>

* Fixed failing unit tests

Signed-off-by: Freya Gustavsson <freya@venefilyn.se>

* Added log phase to logger task output

Signed-off-by: Freya Gustavsson <freya@venefilyn.se>

* chore: remove explicit usages of stages in logs

This removes all the different task log calls that include the stage
it should be in and instead relies on the logger to take care of it

e.g. `.task("Convert: ...")` will now have `Convert:` removed

Signed-off-by: Freya Gustavsson <freya@venefilyn.se>

* chore: Move phases to prevent circular import

ConversionPhases is now moved outside of utils module so that it doesn't
cause a circular import with logger module

Signed-off-by: Freya Gustavsson <freya@venefilyn.se>

* Removed convert2rhel logger from conftest

Signed-off-by: Freya Gustavsson <freya@venefilyn.se>

* Set rollback phase on rollback

Signed-off-by: Freya Gustavsson <freya@spytec.se>
Signed-off-by: Freya Gustavsson <freya@venefilyn.se>

* Log phases correctly

Signed-off-by: Freya Gustavsson <freya@venefilyn.se>

* Set log_name to Prepare for pre PONR changes

Signed-off-by: Freya Gustavsson <freya@venefilyn.se>

* Remove leftover text task stage

* feat: add phase name for analyze

* fix: changed post PONR phase to 'final'

* fix: handle rollback phases in messages

We didn't know what phase we were on before going into rollback. With
this change it now keeps a (soft) linear history where each phase points
to the previous one

* fix: Correctly get the stage we were at

* fix: Update phase from Final to Convert

---------

Signed-off-by: Freya Gustavsson <freya@venefilyn.se>
Signed-off-by: Freya Gustavsson <freya@spytec.se>