Remove legacy/RHEL 5 GPG key
Remove the legacy GPG public key that was used to verify RHEL 5 packages
that were signed with the private part of the key.
Convert2rhel no longer allows converting to RHEL 5 hence there's no need
to bundle the key in the convert2rhel rpm and import it during the
conversion.
Red Hat is using a different key ("release key 2") to sign packages on
RHEL 6/7/8/9. Convert2rhel has been bundling and importing this release
key 2 for a long time.